# quizzard-api/docs/open-api-docs.yaml
openapi: 3.0.1
info:
  title: quizzard
  version: 1.0.11
servers:
  - url: http://localhost:8081
    description: Generated server url
paths:
...
  /quiz:
    post:
      tags:
        - quiz-controller
      summary: Create a new quiz
      operationId: createQuiz
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRequest'
        required: true
      responses:
        '201':
          description: Quiz created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QuizDto'
        '401':
          description: The unauthorized request was sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                unauthorized:
                  description: unauthorized
                  value:
                    errors:
                      - type: /unauthorized
                        title: Unauthorized
                        status: 401
                        detail: The request lacks valid authentication credentials for the requested resource
        '403':
          description: Current user is not authorised for requested action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                access-forbidden:
                  description: access-forbidden
                  value:
                    errors:
                      - type: /access-forbidden
                        title: Access forbidden
                        status: 403
                        detail: The user has no permissions to view the requested resource
        '500':
          description: Unhandled exception occurred during request processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                internal-server-error:
                  description: internal-server-error
                  value:
                    errors:
                      - type: /internal-server-error
                        title: Unexpected internal server error
                        status: 500
                        detail: Unexpected internal server error occurred. Please contact service administrator for more details.
                external-server-error:
                  description: external-server-error
                  value:
                    errors:
                      - type: /external-server-error
                        title: Unexpected external error
                        status: 500
                        detail: Unexpected external server error occurred. Please contact service administrator for more details.
                        source: CommerceTools
      security:
        - bearer-key: []
  /quiz/result:
    post:
      tags:
        - quiz-controller
      summary: Submit quiz result
      operationId: submitResult
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizResult'
        required: true
      responses:
        '201':
          description: Quiz result submitted successfully
        '401':
          description: The unauthorized request was sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                unauthorized:
                  description: unauthorized
                  value:
                    errors:
                      - type: /unauthorized
                        title: Unauthorized
                        status: 401
                        detail: The request lacks valid authentication credentials for the requested resource
        '403':
          description: Current user is not authorised for requested action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                access-forbidden:
                  description: access-forbidden
                  value:
                    errors:
                      - type: /access-forbidden
                        title: Access forbidden
                        status: 403
                        detail: The user has no permissions to view the requested resource
        '500':
          description: Unhandled exception occurred during request processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                internal-server-error:
                  description: internal-server-error
                  value:
                    errors:
                      - type: /internal-server-error
                        title: Unexpected internal server error
                        status: 500
                        detail: Unexpected internal server error occurred. Please contact service administrator for more details.
                external-server-error:
                  description: external-server-error
                  value:
                    errors:
                      - type: /external-server-error
                        title: Unexpected external error
                        status: 500
                        detail: Unexpected external server error occurred. Please contact service administrator for more details.
                        source: CommerceTools
      security:
        - bearer-key: []
...
  /quiz/questions/{category}:
    get:
      tags:
        - quiz-controller
      summary: Get random sequence quiz by category
      operationId: getRandomSequenceQuiz
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Quiz retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QuizDto'
        '401':
          description: The unauthorized request was sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                unauthorized:
                  description: unauthorized
                  value:
                    errors:
                      - type: /unauthorized
                        title: Unauthorized
                        status: 401
                        detail: The request lacks valid authentication credentials for the requested resource
        '403':
          description: Current user is not authorised for requested action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                access-forbidden:
                  description: access-forbidden
                  value:
                    errors:
                      - type: /access-forbidden
                        title: Access forbidden
                        status: 403
                        detail: The user has no permissions to view the requested resource
        '404':
          description: Quiz not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QuizDto'
        '500':
          description: Unhandled exception occurred during request processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                internal-server-error:
                  description: internal-server-error
                  value:
                    errors:
                      - type: /internal-server-error
                        title: Unexpected internal server error
                        status: 500
                        detail: Unexpected internal server error occurred. Please contact service administrator for more details.
                external-server-error:
                  description: external-server-error
                  value:
                    errors:
                      - type: /external-server-error
                        title: Unexpected external error
                        status: 500
                        detail: Unexpected external server error occurred. Please contact service administrator for more details.
                        source: CommerceTools
      security:
        - bearer-key: []
  /quiz/categories:
    get:
      tags:
        - quiz-controller
      summary: Get all categories
      operationId: getCategories
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            '*/*':
              schema:
                type: string
        '401':
          description: The unauthorized request was sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                unauthorized:
                  description: unauthorized
                  value:
                    errors:
                      - type: /unauthorized
                        title: Unauthorized
                        status: 401
                        detail: The request lacks valid authentication credentials for the requested resource
        '403':
          description: Current user is not authorised for requested action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                access-forbidden:
                  description: access-forbidden
                  value:
                    errors:
                      - type: /access-forbidden
                        title: Access forbidden
                        status: 403
                        detail: The user has no permissions to view the requested resource
        '500':
          description: Unhandled exception occurred during request processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                internal-server-error:
                  description: internal-server-error
                  value:
                    errors:
                      - type: /internal-server-error
                        title: Unexpected internal server error
                        status: 500
                        detail: Unexpected internal server error occurred. Please contact service administrator for more details.
                external-server-error:
                  description: external-server-error
                  value:
                    errors:
                      - type: /external-server-error
                        title: Unexpected external error
                        status: 500
                        detail: Unexpected external server error occurred. Please contact service administrator for more details.
                        source: CommerceTools
      security:
        - bearer-key: []
  /quiz/categories/{category}:
    get:
      tags:
        - quiz-controller
      summary: Get categories by prefix
      operationId: getCategoriesByPrefix
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            '*/*':
              schema:
                type: string
        '401':
          description: The unauthorized request was sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                unauthorized:
                  description: unauthorized
                  value:
                    errors:
                      - type: /unauthorized
                        title: Unauthorized
                        status: 401
                        detail: The request lacks valid authentication credentials for the requested resource
        '403':
          description: Current user is not authorised for requested action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                access-forbidden:
                  description: access-forbidden
                  value:
                    errors:
                      - type: /access-forbidden
                        title: Access forbidden
                        status: 403
                        detail: The user has no permissions to view the requested resource
        '500':
          description: Unhandled exception occurred during request processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                internal-server-error:
                  description: internal-server-error
                  value:
                    errors:
                      - type: /internal-server-error
                        title: Unexpected internal server error
                        status: 500
                        detail: Unexpected internal server error occurred. Please contact service administrator for more details.
                external-server-error:
                  description: external-server-error
                  value:
                    errors:
                      - type: /external-server-error
                        title: Unexpected external error
                        status: 500
                        detail: Unexpected external server error occurred. Please contact service administrator for more details.
                        source: CommerceTools
      security:
        - bearer-key: []
...
components:
  schemas:
    ApiErrorResponse:
      required:
        - errors
      type: object
      properties:
        errors:
          required:
            - type
          type: array
          description: The array of occurred errors during request execution
          items:
            $ref: '#/components/schemas/Problem'
      description: The wrapper of array of occurred errors during request execution
    Problem:
      required:
        - type
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: '#/components/schemas/StatusType'
        detail:
          type: string
        title:
          type: string
      description: The array of occurred errors during request execution
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
...
    QuizRequest:
      type: object
      properties:
        category:
          type: string
        questionText:
          type: string
        correctAnswer:
          type: string
        possibleAnswers:
          type: array
          items:
            type: string
    QuestionDto:
      type: object
      properties:
        questionText:
          type: string
        possibleAnswers:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        questionPicture:
          type: array
          items:
            type: string
            format: byte
    QuizDto:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionDto'
    QuizResult:
      type: object
      properties:
        category:
          type: string
        playerScores:
          type: object
          additionalProperties:
            type: integer
            format: int32
...
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT